some html layout tweaks

+++ src/html.rs
@@ -2,6 +2,14 @@ use std::fs::File;
 use std::io::{self, Write};
 // use std::path::Path;
 
+const BODY_FONT_FAMILY: &str = "monospace";
+const CLASS_HEAD_POSITION: &str = "head_pos";
+const CLASS_CHANGED_POSITION: &str = "change_pos";
+const CSS_HEAD_POSITION_LIGHT: &str =
+    "color: black; background-color: lavenderblush; font-weight: bold;";
+const CSS_HEAD_POSITION_DARK: &str = "color: black; background-color: lavenderblush;";
+const CSS_CHANGE_POSITION_LIGHT: &str = "color: blue; font-weight: bold;";
+
 pub fn create_html() -> io::Result<()> {
     // Define file names
     let html_file_name = "index.html";
@@ -13,12 +21,16 @@ pub fn create_html() -> io::Result<()> {
     write_html_content(&mut html_file, light_css_file_name, dark_css_file_name)?;
 
     // Create and write to light mode CSS file
-    let mut light_css_file = File::create(light_css_file_name)?;
-    write_light_css_content(&mut light_css_file)?;
+    if !file_exists(light_css_file_name) {
+        let mut light_css_file = File::create(light_css_file_name)?;
+        write_light_css_content(&mut light_css_file)?;
+    }
 
     // Create and write to dark mode CSS file
-    let mut dark_css_file = File::create(dark_css_file_name)?;
-    write_dark_css_content(&mut dark_css_file)?;
+    if !file_exists(dark_css_file_name) {
+        let mut dark_css_file = File::create(dark_css_file_name)?;
+        write_dark_css_content(&mut dark_css_file)?;
+    }
 
     println!(
         "Successfully created '{}', '{}', and '{}'",
@@ -58,22 +70,23 @@ fn write_html_content(file: &mut File, light_css: &str, dark_css: &str) -> io::R
         light_css
     )?; // Fallback for browsers not supporting prefers-color-scheme
     writeln!(file, "    <style>")?;
-    writeln!(file, "        body {{ font-family: sans-serif; }}")?;
     writeln!(
         file,
-        "        .blue-two {{ color: blue; background-color: gray; padding: 2px 0; }}"
+        "        body {{ font-family: {BODY_FONT_FAMILY}; font-size: larger;}}"
     )?;
     writeln!(file, "    </style>")?;
     writeln!(file, "</head>")?;
     writeln!(file, "<body>")?;
 
+    writeln!(file, "<p>")?;
     for i in 1..=100 {
         writeln!(
             file,
-            "    <p>{}: 1RB 001<span class=\"blue-two\">2</span>0</p>",
-            i
+            "{}: 1RB 001<span class=\"{CLASS_HEAD_POSITION}\">2</span><span class=\"{CLASS_CHANGED_POSITION}\">0</span></br>",
+            format_int(i, 10),
         )?;
     }
+    writeln!(file, "</p>")?;
 
     writeln!(file, "</body>")?;
     writeln!(file, "</html>")?;
@@ -85,6 +98,15 @@ fn write_light_css_content(file: &mut File) -> io::Result<()> {
     writeln!(file, "    background-color: white;")?;
     writeln!(file, "    color: black;")?;
     writeln!(file, "}}")?;
+    writeln!(file)?;
+    writeln!(
+        file,
+        ".{CLASS_HEAD_POSITION} {{\n    {CSS_HEAD_POSITION_LIGHT}\n}}"
+    )?;
+    writeln!(
+        file,
+        ".{CLASS_CHANGED_POSITION} {{\n    {CSS_CHANGE_POSITION_LIGHT}\n}}"
+    )?;
     Ok(())
 }
 
@@ -93,5 +115,17 @@ fn write_dark_css_content(file: &mut File) -> io::Result<()> {
     writeln!(file, "    background-color: black;")?;
     writeln!(file, "    color: white;")?;
     writeln!(file, "}}")?;
+    writeln!(file)?;
+    writeln!(file, ".{CLASS_HEAD_POSITION} {{{CSS_HEAD_POSITION_DARK}}}")?;
     Ok(())
+}
+
+// check if a file exists
+fn file_exists(file_path: &str) -> bool {
+    std::path::Path::new(file_path).exists()
+}
+
+fn format_int(number: usize, size: usize) -> String {
+    let mut s = format!("{:>size$}", number);
+    s.replace(" ", "&nbsp;")
 }
